Refactored my server and modularized the functions and classes into two separate JS files for the weather and the movies and put them in a modules folder. Practiced refactoring the try / catch in my Movie module JS file. Update the README.
